Extract the Scout runtime out of setup
